requirements.txt
dill==0.3.5.1
dnspython==2.2.1
fastapi==0.78.0
httptools==0.5.0
imblearn==0.0
mypy-boto3-s3==1.24.76
pip-chill==1.0.1
pymongo==4.2.0
python-dotenv==0.21.0
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==1.6.2
pandas~=1.5.2
PyYAML~=6.0
numpy~=1.23.4
scikit-learn~=1.1.3
apache-airflow
pip~=22.3.1
attrs~=22.2.0
colorlog~=4.8.0
zlib~=1.2.13
wheel~=0.37.1
openssl~=1.1.1s
cryptography~=38.0.1
docutils~=0.19
jinja2~=3.1.2
setuptools~=65.5.0
sensor~=0.0.1
scipy~=1.9.3
yaml~=0.2.5
sklearn~=0.0.post1
pendulum~=2.1.2
-e .